Sprint 4 - Fixed bullet
Changed bullet so it works better. Added a new node called "EnemyBulletDirection" under the player node in the game scene. The enemy bullet finds that node and travels to that node rather than traveling to the player. This helped the bullet travel to the player rather than it's feet. Added in a bullet delay function to enable the use of attack cycles. Fixed the bullet rotation so it faces towards the player rather than away from it. Changed bullet speed from 250 to 270.